Added functionality for matching items with desired terms.
Wanted terms can be specified on each item in the `item_attributes` list, like so:
      - Breed: '<span class="breed">([^<]+)</span>'
        want:
          - Snowshoe
          - Siamese

Another term, on the top level, specifies whether to match an item loosely (if any of its `want`s match) or strictly (only if all its `want`s do).
  wants: "all"

The only value that has any meaning for this key is "all". And, since the default mode is loose, really it should only be added if strict matching is desired.